Move ConfirmDialog handling to useConfirmDialog hook
