+++ oarepo_requests/ui/theme/assets/semantic-ui/js/oarepo_requests_ui/record-requests/components/RequestModal.jsx
@@ -12,6 +12,7 @@ import axios from "axios";
 import { RequestModalContent, CreateRequestModalContent } from ".";
 import { REQUEST_TYPE } from "../utils/objects";
 import { isDeepEmpty, mapPayloadUiToInitialValues } from "../utils";
+import { useConfirmDialog } from "../utils/hooks";
 
 /** 
  * @typedef {import("../types").Request} Request
@@ -26,22 +27,14 @@ export const RequestModal = ({ request, requestTypes, requestModalType, isEventM
   const [modalOpen, setModalOpen] = useState(false);
   const [error, setError] = useState(null);
 
-  /** @type {[ConfirmProps, (props: ConfirmProps) => void]} */
-  const [confirmDialogProps, setConfirmDialogProps] = useState({
-    open: false,
-    content: i18next.t("Are you sure?"),
-    cancelButton: i18next.t("Close"),
-    confirmButton: i18next.t("OK"),
-    onCancel: () => setConfirmDialogProps(props => ({ ...props, open: false })),
-    onConfirm: () => setConfirmDialogProps(props => ({ ...props, open: false }))
-  });
-
   const errorMessageRef = useRef(null);
 
   const formik = useFormik({
     initialValues: !_isEmpty(request?.payload) ? { payload: request.payload } : (request?.payload_ui ? mapPayloadUiToInitialValues(request?.payload_ui) : {}),
     onSubmit: () => {}
   });
+
+  const { confirmDialogProps, confirmAction } = useConfirmDialog(formik, sendRequest, isEventModal);
 
   useEffect(() => {
     if (error) {
@@ -109,59 +102,6 @@ export const RequestModal = ({ request, requestTypes, requestModalType, isEventM
     const mappedData = !isDeepEmpty(formik.values) ? {} : formik.values;
     const actionUrl = !isEventModal ? request.links.actions[requestType] : request.links[requestType];
     return callApi(actionUrl, 'post', mappedData);
-  }
-
-  const confirmAction = (requestType, createAndSubmit = false) => {
-    /** @type {ConfirmProps} */
-    let newConfirmDialogProps = {
-      open: true,
-      onConfirm: () => {
-        setConfirmDialogProps(props => ({ ...props, open: false }));
-        sendRequest(requestType);
-      },
-      onCancel: () => {
-        setConfirmDialogProps(props => ({ ...props, open: false }));
-        formik.setSubmitting(false);
-      }
-    };
-
-    switch (requestType) {
-      case REQUEST_TYPE.CREATE:
-        newConfirmDialogProps.header = isEventModal ? i18next.t("Submit event") : i18next.t("Create request");
-        break;
-      case REQUEST_TYPE.SUBMIT:
-        newConfirmDialogProps.header = i18next.t("Submit request");
-        newConfirmDialogProps.confirmButton = i18next.t("OK");
-        break;
-      case REQUEST_TYPE.CANCEL:
-        newConfirmDialogProps.header = i18next.t("Cancel request");
-        newConfirmDialogProps.confirmButton = <Button negative>{i18next.t("Cancel request")}</Button>;
-        break;
-      case REQUEST_TYPE.ACCEPT:
-        newConfirmDialogProps.header = i18next.t("Accept request");
-        newConfirmDialogProps.confirmButton = <Button positive>{i18next.t("Accept")}</Button>;
-        break;
-      case REQUEST_TYPE.DECLINE:
-        newConfirmDialogProps.header = i18next.t("Decline request");
-        newConfirmDialogProps.confirmButton = <Button negative>{i18next.t("Decline")}</Button>;
-        break;
-      default:
-        break;
-    }
-
-    if (createAndSubmit) {
-      newConfirmDialogProps = {
-        ...newConfirmDialogProps,
-        header: i18next.t("Create and submit request"),
-        confirmButton: <Button positive>{i18next.t("Create and submit")}</Button>,
-        onConfirm: () => {
-          setConfirmDialogProps(props => ({ ...props, open: false }));
-          sendRequest(REQUEST_TYPE.CREATE, createAndSubmit);
-        }
-      }
-    }
-
-    setConfirmDialogProps(props => ({ ...props, ...newConfirmDialogProps }));
   }
 
   const onClose = () => {
